+++ src/pages/Admin/HomeAdmin.jsx
@@ -27,6 +27,7 @@ import "react-calendar/dist/Calendar.css";
 import ".././Home.css";
 import { useNavigate } from "react-router-dom";
 import Swal from "sweetalert2";
+import { swalWithBootstrapButtons } from "../../hooks/swalWithBootstrapButtons";
 
 const Home = () => {
   const user = useSelector((state) => state.users);
@@ -140,10 +141,27 @@ const Home = () => {
     }
   };
 
-  const cancelTurn = async (id) => {
-    const turnDoc = doc(db, "turnos", id);
-    await deleteDoc(turnDoc);
-    getTurns();
+  const cancelTurn = (id) => {
+    swalWithBootstrapButtons
+      .fire({
+        title: "Estas seguro?",
+        text: "No podrás volver a recuperar esta acción!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonText: "Si, eliminar turno!",
+      })
+      .then(async (result) => {
+        if (result.isConfirmed) {
+          const turnDoc = doc(db, "turnos", id);
+          await deleteDoc(turnDoc);
+          swalWithBootstrapButtons.fire(
+            "Turno eliminado!",
+            "El turno ha sido eliminado.",
+            "success"
+          );
+        }
+        getTurns();
+      });
   };
 
   return (
